refactor: move UserDomainEntity to its own file
